+++ packages/layout/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.5](https://github.com/medly/medly-components/compare/@medly-components/layout@1.23.4...@medly-components/layout@1.23.5) (2021-07-28)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.23.4](https://github.com/medly/medly-components/compare/@medly-components/layout@1.23.3...@medly-components/layout@1.23.4) (2021-07-27)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.23.3](https://github.com/medly/medly-components/compare/@medly-components/layout@1.23.2...@medly-components/layout@1.23.3) (2021-07-27)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.23.2](https://github.com/medly/medly-components/compare/@medly-components/layout@1.23.1...@medly-components/layout@1.23.2) (2021-07-26)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
 ## [1.23.1](https://github.com/medly/medly-components/compare/@medly-components/layout@1.23.0...@medly-components/layout@1.23.1) (2021-07-19)
 
 **Note:** Version bump only for package @medly-components/layout